requirements.txt
# Core Dependencies
flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
requests==2.31.0

# AI/ML
openai==1.3.0
anthropic==0.7.0

# Video Generation
# d-id (uses requests, no specific package)

# Faceless Video Dependencies
playwright==1.40.0
Pillow==10.1.0
matplotlib==3.8.2
numpy==1.24.3
pyttsx3==2.90  # Fallback TTS

# Audio Processing
elevenlabs==0.2.26

# Async Support
asyncio==3.4.3
aiohttp==3.9.1

# Web Scraping & Research
beautifulsoup4==4.12.2
selenium==4.15.2

# Data Processing
pandas==2.0.3
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Performance
redis==5.0.1
cachetools==5.3.2

# Utilities
pydantic==2.5.0
typing-extensions==4.8.0